requirements.txt
requests
beautifulsoup4
lxml
jsonschema
progressbar2